Solved PR issues
