--- routes/randomstring.js
@@ -1,7 +1,0 @@
-function randomString(length, chars){
-    let result = '';
-    for (let i = length; i > 0; --i) result += chars[Math.floor(Math.random() * chars.length)];
-    return result;
-}
-
-module.exports=randomString;
