+++ src/_variables.scss
@@ -1,0 +1,3 @@
+$header-font-weight: 900;
+$header-font-size: 200%;
+$error-message-color: #ED4337;
