+++ src/app/splashscreen/splashscreen.component.html
@@ -1,5 +1,4 @@
-
-<div style="text-align:center">
+<div class="text-center">
 	<H1><span class="titleleft">PC Prep</span><span class="titleright">Kit</span></H1>
 	<div >
 	<img src="assets/img/logo.png" class ="logo">
@@ -15,5 +14,4 @@
   		<span class="sr-only">Progress Bar</span>
   		</div>
   	</div>
-
 </div>
