+++ src/app/splashscreen/splashscreen.component.ts
@@ -1,35 +1,31 @@
 import { Component, OnInit } from '@angular/core';
-import {Router} from "@angular/router";
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-splashscreen',
   templateUrl: './splashscreen.component.html',
   styleUrls: ['./splashscreen.component.scss']
 })
-export class SplashscreenComponent implements OnInit {
+export class SplashscreenComponent {
 
-  wid:number=0;
+  wid = 0 ;
   intervalId;
   constructor(private router: Router) {
-  		this.start();
-  	}
-
-
-  start(){
-  this.intervalId=setInterval(()=>{
-  		 this.wid=this.wid+5;
-  		 if(this.wid>100){
-  		 	this.stop();
-  		 }
-  		 }, 70);
+    this.start();
   }
 
-  stop(){
-  	clearInterval(this.intervalId);
-  	this.router.navigate(['login']);
-
+  start() {
+      this.intervalId = setInterval(() => {
+      this.wid = this.wid + 5 ;
+      if ( this.wid > 100) {
+        this.stop();
+      }
+    }, 70);
   }
-  ngOnInit() {
+
+  stop() {
+    clearInterval(this.intervalId);
+    this.router.navigate(['login']);
   }
 
 }
